+++ src/main/java/com/braintreegateway/OAuthGateway.java
@@ -39,10 +39,10 @@ public class OAuthGateway {
     public String connectUrl(OAuthConnectUrlRequest request) {
         request.clientId(configuration.getClientId());
         String queryString = request.toQueryString();
-        String url = configuration.getBaseURL() + "/oauth/connect?" + queryString;
-        return String.format("%1$s&signature=%2$s&algorithm=SHA256", url, computeSignature(url));
+        return configuration.getBaseURL() + "/oauth/connect?" + queryString;
     }
 
+    @Deprecated
     public String computeSignature(String url) {
         Sha256Hasher hasher = new Sha256Hasher();
         return hasher.hmacHash(configuration.getClientSecret(), url);