+++ src/main/java/com/braintreegateway/RiskData.java
@@ -6,10 +6,13 @@ public class RiskData {
 
     private String id;
     private String decision;
+    private Boolean deviceDataCaptured;
 
     public RiskData(NodeWrapper node) {
         id = node.findString("id");
         decision = node.findString("decision");
+        String deviceDataCapturedString = node.findString("device-data-captured");
+        deviceDataCaptured = (deviceDataCapturedString == null) ? null : Boolean.valueOf(deviceDataCapturedString);
     }
 
     public String getId() {
@@ -18,5 +21,9 @@ public class RiskData {
 
     public String getDecision() {
         return decision;
+    }
+
+    public Boolean getDeviceDataCaptured() {
+        return deviceDataCaptured;
     }
 }