+++ src/main/java/com/braintreegateway/util/EnumUtils.java
@@ -12,4 +12,16 @@ public class EnumUtils {
             return defaultValue;
         }
     }
+
+    public static <T extends Enum<T>> T findByToString(T[] values, String name, T defaultValue) {
+        if (name == null || values == null) {
+            return defaultValue;
+        }
+        for (T value : values) {
+            if (name.equals(value.toString())) {
+                return value;
+            }
+        }
+        return defaultValue;
+    }
 }