+++ src/test/java/com/braintreegateway/integrationtest/OAuthIT.java
@@ -102,6 +102,7 @@ public class OAuthIT extends IntegrationTest {
             scope("read_write").
             state("baz_state").
             landingPage("login").
+            loginOnly(true).
             user().
                 country("USA").
                 email("foo@example.com").
@@ -153,6 +154,7 @@ public class OAuthIT extends IntegrationTest {
             assertEquals("read_write", query.get("scope"));
             assertEquals("baz_state", query.get("state"));
             assertEquals("login", query.get("landing_page"));
+            assertEquals("true", query.get("login_only"));
 
             assertEquals("USA", query.get("user[country]"));
 
@@ -186,11 +188,6 @@ public class OAuthIT extends IntegrationTest {
             assertEquals("USD", query.get("business[currency]"));
             assertEquals("http://example.com", query.get("business[website]"));
             assertEquals("1988-10", query.get("business[established_on]"));
-
-            assertEquals(64, query.get("signature").length());
-            assertTrue(query.get("signature").matches("^[a-f0-9]+$"));
-            assertEquals("SHA256", query.get("algorithm"));
-
         } catch (java.io.UnsupportedEncodingException e) {
             fail("unsupported encoding");
         } catch (java.net.MalformedURLException e) {
@@ -242,6 +239,7 @@ public class OAuthIT extends IntegrationTest {
         }
     }
 
+    @SuppressWarnings("deprecation")
     @Test
     public void computeSignatureReturnsCorrectSignature() {
         String url = "http://localhost:3000/oauth/connect?business%5Bname%5D=We+Like+Spaces&client_id=client_id%24development%24integration_client_id";