+++ src/test/java/com/braintreegateway/util/EnumUtilsTest.java
@@ -1,13 +1,24 @@
 package com.braintreegateway.util;
 
+import com.braintreegateway.CreditCard;
 import com.braintreegateway.Transaction;
-import com.braintreegateway.Dispute;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
 
 public class EnumUtilsTest {
+
+    public enum TestEnum {
+        VALUE1,
+        VALUE2,
+        VALUE3;
+
+        @Override
+        public String toString() {
+            return null;
+        }
+    }
 
     @Test
     public void findByNameWithNull() {
@@ -28,5 +39,40 @@ public class EnumUtilsTest {
     public void findByNameDefaultsToProvidedDefaultIfNameDoesNotMatch() {
         assertEquals(Transaction.Status.UNRECOGNIZED, EnumUtils.findByName(Transaction.Status.class, "blah", Transaction.Status.UNRECOGNIZED));
         assertEquals(Transaction.Type.UNRECOGNIZED, EnumUtils.findByName(Transaction.Type.class, "blah", Transaction.Type.UNRECOGNIZED));
+    }
+
+    @Test
+    public void findByToString_returnsEnumWithExactMatch() {
+        assertEquals(CreditCard.Commercial.NO, EnumUtils.findByToString(CreditCard.Commercial.values(), "No", CreditCard.Commercial.UNKNOWN));
+    }
+
+    @Test
+    public void findByToString_returnsDefaultIfNullName() {
+        assertEquals(CreditCard.Commercial.UNKNOWN, EnumUtils.findByToString(CreditCard.Commercial.values(), null, CreditCard.Commercial.UNKNOWN));
+    }
+
+    @Test
+    public void findByToString_returnsDefaultIfNullValues() {
+        assertEquals(CreditCard.Commercial.UNKNOWN, EnumUtils.findByToString(null, "No", CreditCard.Commercial.UNKNOWN));
+    }
+
+    @Test
+    public void findByToString_returnsDefaultIfCaseNotAMatch() {
+        assertEquals(CreditCard.Commercial.UNKNOWN, EnumUtils.findByToString(CreditCard.Commercial.values(), "NO", CreditCard.Commercial.UNKNOWN));
+    }
+
+    @Test
+    public void findByToString_returnsDefaultIfNotFound() {
+        assertEquals(CreditCard.Commercial.UNKNOWN, EnumUtils.findByToString(CreditCard.Commercial.values(), "404", CreditCard.Commercial.UNKNOWN));
+    }
+
+    @Test
+    public void findByToString_returnsNullIfNotAMatchAndDefaultValueIsNull() {
+        assertNull(EnumUtils.findByToString(CreditCard.Commercial.values(), "404", null));
+    }
+
+    @Test
+    public void findByToString_returnsDefaultIfToStringIsNull() {
+        assertEquals(TestEnum.VALUE3, EnumUtils.findByToString(TestEnum.values(), "VALUE1", TestEnum.VALUE3));
     }
 }